+++ src/pages/Home/HomePage.tsx
@@ -5,74 +5,22 @@ import { BottomSheetContent } from '@/components/HomePage/BottomSheet/BottomShee
 import { LetterContainer } from '@/components/HomePage/LetterContainer/LetterContainer';
 import { WelcomeMessageContainer } from '@/components/HomePage/WelcomeMessageContainer/WelcomeMessageContainer';
 import { useUserStore } from '@/stores/useUserStore';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { TopButtonContainer } from '@/components/HomePage/TopButtonContainer/TopButtonContainer';
-import { getToken, firebaseMessaging, onMessage } from '@/util/firebase';
-import { postToken } from '@/service/nofication/postToken';
-import { useToastStore } from '@/hooks';
 import { Loading } from '@/components/Common/Loading/Loading';
 import { ToggleVariant } from '@/constants/toggleVariant';
 import { useGetThreeLetterData } from '@/hooks/useGetThreeLetterData';
 import { ErrorPage } from '../ErrorPage';
+import { usePushNotifications } from '@/hooks/usePushNotifications';
 
 export const HomePage = () => {
-    const { addToast } = useToastStore();
-
-    const isToken = localStorage.getItem('isToken');
-
-    useEffect(() => {
-        if (isToken === 'true') {
-            return;
-        }
-        const handlePushNotifications = async () => {
-            try {
-                const permission = await Notification.requestPermission();
-
-                if (permission === 'granted') {
-                    const token = await getToken(firebaseMessaging, {
-                        vapidKey: import.meta.env.VITE_FCM_VAPID_KEY
-                    });
-
-                    try {
-                        const data = await postToken({ token });
-                        addToast(`${data.message}`, 'success');
-                        localStorage.setItem('isToken', 'true');
-                    } catch (error) {
-                        console.error(error);
-                    }
-                } else {
-                    console.log('Notification permission denied');
-                }
-            } catch (error) {
-                console.error('FCM Error : ', error);
-            }
-        };
-
-        // 푸시 알림 설정 및 수신 처리
-        handlePushNotifications();
-
-        // 푸시 알림 메시지 수신 시 처리
-        const unsubscribe = onMessage(firebaseMessaging, (payload) => {
-            if (payload.notification) {
-                alert(payload.notification.body);
-            }
-        });
-
-        // 컴포넌트가 언마운트 될 때 리스너 해제
-        return () => {
-            unsubscribe();
-        };
-    }, []); // 빈 배열을 의존성으로 사용해 처음 마운트될 때만 실행되게 함
+    usePushNotifications(); // 푸시 알림 훅
 
     const { user } = useUserStore();
     const [open, setOpen] = useState(false);
     const [toggle, setToggle] = useState(true);
 
     const { letters, isLoading, isError } = useGetThreeLetterData(toggle);
-
-    function onDismiss() {
-        setOpen(false);
-    }
 
     if (isLoading) {
         return <Loading />;
@@ -120,7 +68,7 @@ export const HomePage = () => {
 
                 <SliderMenuContainer
                     open={open}
-                    onDismiss={onDismiss}
+                    onDismiss={() => setOpen(false)}
                     snapPoints={() => [window.innerHeight * 0.95]}
                 >
                     <BottomSheetContent
